Rework example server - new loopback api, android
The server mounts REST endpoint at /api, which is the setup required
by the recent versions of loopback.

The custom route for sending notifications is mounted as /notify/:id
(outside of /api space).

The server does not overwrite an already registered application. This
way the devices already registered are preserved across restarts.
Also since we are running on a database publicly shared by the whole
world, the users won't override their settings now.

Last but not least: add GCM configuration, so that push notification can
be sent to Android devices too. Note the GCM provider is not completely
implemented yet, there are some edge cases where an error is thrown.