update module to loopback 2.0
The existing code had models that used beforeCreate calls which can be
replaced by the operation hooks.  My current code is using loopback 2.0
so importing this was causing errors.

Here’s the breakdown:

* Created a Grunfile to help with the local doc generation
* Added an index.js file to the models directory to load up the models
from both JSON and JS files.  Mostly borrowed from
https://github.com/strongloop/loopback-component-passport/blob/master/li
b/index.js
* In the push-manager I moved from `getModel` to `findModel` because
7316048afc1164ed91a004a8f4f58f3603df456f changed `getModel` to throw an
error.
* The Installation model was stripped down to do most of the work in
the JSON file and remove the `beforeCreate` call in favour of a ‘before
save’ observer.  The `created` attribute is now generated from the
`defaultFn`.  And the `modified` attribute can safely be updated in the
‘before save’ function.
* The Notification model was stripped down so the definition was in the
JSON file.  A lot of the comments within are duplicated though I did
like the inline comments more than the top level jsdoc ones.
Similarly the `beforeCreate` function was replaced with a ‘before save’
function.
* package.json: npm sorted the dependencies which makes the diff
annoying.
 * I updated the debug module, couldn’t see why we needed to stay on
the 1.x version
 * added lodash as I used it in a couple places and intend to use it a
bit more as I clean up more of the code in general
 * updated loopback to 2.x
 * updated loopback-datasource-juggler to 2.x as well
 * updated loopback-testing to 1.1.x
 * simplified the test script with the help of moca.opts

Tests
* I added a common.js file which I haven’t seen this team use yet but
have seen as common practice in other places so I don’t have repeat
myself with lots of includes on every test.  I’m inclined to move most
of the testings across modules to this kind of format if this approach
is acceptable.
* Added a mocha.opts file which takes the options away from the
package.json
* Thanks to the common.js I removed a lot of duplicate includes from
the test files.